testing theater mode
